\name{BPterrainAnal-package}
\alias{BPterrainAnal-package}
\alias{BPterrainAnal}
\docType{package}
\title{
Uses different forms of terrain analyses.
}
\description{
It contains all the function used in the main script of the applied geo-scripting assignment.
}
\details{
\tabular{ll}{
Package: \tab BPterrainAnal\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-06\cr
License: \tab Open Source\cr
}

}
\author{
B.P. Ottow

<boukepieter.ottow@wur.nl>
}
\references{
http://grass.osgeo.org/
}
\keyword{ package }
\keyword{ dem }
\keyword{ terrain-analysis }
\keyword{ hydrology }
\keyword{ GRASS }
\keyword{ slope }
\seealso{


}
\examples{
# see main script for examples.
}
